Extract sanitizeBlockAttributes and use it on both createBlock and cloneBlock
